feat(jwt): create new module to organize jwt functions
