+++ src/main.rs
@@ -9,6 +9,7 @@ extern crate rocket;
 
 mod database;
 mod graphql_resolvers;
+mod jwt;
 mod models;
 mod schema;
 mod sendemail;
