+++ src/services/user.rs
@@ -3,12 +3,9 @@ use std::fmt;
 use chrono::{NaiveDateTime, Utc};
 use rand::Rng;
 
-use hmac::{Hmac, Mac};
-use jwt::{SignWithKey, VerifyWithKey};
-use serde::{Deserialize, Serialize};
-use sha2::Sha256;
 use uuid::Uuid;
 
+use crate::jwt;
 use crate::models::user;
 use crate::sendemail::send_code;
 
@@ -17,9 +14,7 @@ use crate::database;
 #[derive(Debug)]
 pub enum UserServiceError {
     UserModelFailed(user::UserModelError),
-    FailedToGenerateJwtKey(sha2::digest::InvalidLength),
-    FailedToGenerateJwtToken(jwt::Error),
-    FailedToVerifyJwtToken(jwt::Error),
+    JwtError(jwt::JwtError),
     LoginCodeNotMatching,
 }
 
@@ -32,6 +27,12 @@ impl fmt::Display for UserServiceError {
 impl From<user::UserModelError> for UserServiceError {
     fn from(error: user::UserModelError) -> Self {
         UserServiceError::UserModelFailed(error)
+    }
+}
+
+impl From<jwt::JwtError> for UserServiceError {
+    fn from(error: jwt::JwtError) -> Self {
+        UserServiceError::JwtError(error)
     }
 }
 
@@ -54,11 +55,6 @@ pub struct NewUser {
     pub email: String,
     pub last_code_gen_request: NaiveDateTime,
     pub login_code: i32,
-}
-
-#[derive(Serialize, Deserialize)]
-struct TokenContent {
-    email: String,
 }
 
 pub fn create_user(
@@ -85,7 +81,7 @@ pub fn delete_user(
     token: String,
     jwt_secret: &str,
 ) -> Result<User, UserServiceError> {
-    let email = verify_token(&token, jwt_secret)?;
+    let email = jwt::verify_token(&token, jwt_secret)?;
     Ok(user::delete_user(conn, email)?)
 }
 
@@ -98,63 +94,8 @@ pub fn validate_and_generate_token(
     let real_login_code = user::get_login_code(conn, &email)?;
 
     if login_code == real_login_code {
-        generate_token(&email, jwt_secret)
+        Ok(jwt::generate_token(&email, jwt_secret)?)
     } else {
         Err(UserServiceError::LoginCodeNotMatching)
-    }
-}
-
-fn generate_token(email: &str, jwt_secret: &str) -> Result<String, UserServiceError> {
-    let key: Hmac<Sha256> = Hmac::new_from_slice(jwt_secret.as_bytes())
-        .map_err(UserServiceError::FailedToGenerateJwtKey)?;
-
-    TokenContent {
-        email: email.to_string(),
-    }
-    .sign_with_key(&key)
-    .map_err(UserServiceError::FailedToGenerateJwtToken)
-}
-
-fn verify_token(token: &str, jwt_secret: &str) -> Result<String, UserServiceError> {
-    let key: Hmac<Sha256> = Hmac::new_from_slice(jwt_secret.as_bytes())
-        .map_err(UserServiceError::FailedToGenerateJwtKey)?;
-    let content: TokenContent = token
-        .verify_with_key(&key)
-        .map_err(UserServiceError::FailedToVerifyJwtToken)?;
-    Ok(content.email)
-}
-
-#[cfg(test)]
-mod jwt_tests {
-    use super::*;
-
-    const SOME_EMAIL: &str = "someemail@gmail.com";
-    const GOOD_SECRET: &str = "1234567890987654321";
-    const BAD_SECRET: &str = "1234567890987654322";
-    const GOOD_TOKEN: &str = "eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InNvbWVlbWFpbEBnbWFpbC5jb20ifQ.8U6HQs0S3UZtbBwUcz2cJwD7d0C6op5QTMsSB1402ys";
-    const BAD_TOKEN: &str = "eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InNvbWVlbWFpbEBnbWFpbC5jb20ifQ.8U6HQs0S3UZtbBwUcz2cJwD7d0C6op5QTMsSB1402yt";
-
-    #[test]
-    fn can_generate_token() {
-        let result = generate_token(SOME_EMAIL, GOOD_SECRET).unwrap();
-        assert_eq!(result, GOOD_TOKEN);
-    }
-
-    #[test]
-    fn can_get_email_back() {
-        let result = verify_token(GOOD_TOKEN, GOOD_SECRET).unwrap();
-        assert_eq!(result, SOME_EMAIL)
-    }
-
-    #[test]
-    fn cant_verify_bad_token() {
-        let result = verify_token(BAD_TOKEN, GOOD_SECRET);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn cant_verify_bad_secret() {
-        let result = verify_token(GOOD_TOKEN, BAD_SECRET);
-        assert!(result.is_err());
     }
 }
